pythonProject/requirements.txt
anyio==4.2.0
APScheduler==3.6.3
bitcoinlib==0.6.13
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
coinpayments-py==0.1.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastecdsa==2.3.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
mysql-connector-python==8.3.0
numpy>=1.19.5
pycoinpayments==1.0.0
pycryptodome==3.20.0
python-telegram-bot==13.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
telegram==0.0.1
tornado==6.4
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
